prediction offset issue fixed
create_training_sequence() was creating a tensor with a number of samples shorter than the ground truth vector which caused the predictions to offset by an amount equal to (sequence_length - 1).

Current solution is truncating the ground truth vector such that the first (sequence_length - 1) values are removed. 
Resetting the index of this truncated vector gives predictions that are temporally aligned.